Moving the TLSConfig to structs
